+++ docs/changelog.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BanSubstrings` scanner to handle substrings with special characters.
 
 ### Changed
--
+- `Gibberish` scanner has higher threshold to reduce false positives. In addition, it supports changing `labels` to remove overtriggering when `mild gibberish` is detected.
 
 ### Removed
 -